Restructure Lock to follow new paradigm
