+++ src/Core/Lock/Type/CacheLock.php
@@ -19,9 +19,8 @@
  *
  */
 
-namespace Friendica\Core\Lock;
+namespace Friendica\Core\Lock\Type;
 
-use Friendica\Core\BaseLock;
 use Friendica\Core\Cache\Enum\Duration;
 use Friendica\Core\Cache\IMemoryCache;
