+++ src/Core/Lock/Type/DatabaseLock.php
@@ -19,10 +19,10 @@
  *
  */
 
-namespace Friendica\Core\Lock;
+namespace Friendica\Core\Lock\Type;
 
-use Friendica\Core\BaseLock;
 use Friendica\Core\Cache\Enum\Duration;
+use Friendica\Core\Lock\Enum\Type;
 use Friendica\Database\Database;
 use Friendica\Util\DateTimeFormat;
