+++ tests/src/Core/Lock/MemcacheCacheLockTest.php
@@ -24,7 +24,7 @@ namespace Friendica\Test\src\Core\Lock;
 use Exception;
 use Friendica\Core\Cache\Type\MemcacheCache;
 use Friendica\Core\Config\IConfig;
-use Friendica\Core\Lock\CacheLock;
+use Friendica\Core\Lock\Type\CacheLock;
 use Mockery;
 
 /**
@@ -53,7 +53,7 @@ class MemcacheCacheLockTest extends LockTest
 
 		try {
 			$cache = new MemcacheCache($host, $configMock);
-			$lock = new CacheLock($cache);
+			$lock = new \Friendica\Core\Lock\Type\CacheLock($cache);
 		} catch (Exception $e) {
 			static::markTestSkipped('Memcache is not available');
 		}
