+++ tests/src/Core/Lock/MemcachedCacheLockTest.php
@@ -24,7 +24,7 @@ namespace Friendica\Test\src\Core\Lock;
 use Exception;
 use Friendica\Core\Cache\Type\MemcachedCache;
 use Friendica\Core\Config\IConfig;
-use Friendica\Core\Lock\CacheLock;
+use Friendica\Core\Lock\Type\CacheLock;
 use Mockery;
 use Psr\Log\NullLogger;
