+++ tests/src/Core/Lock/RedisCacheLockTest.php
@@ -24,7 +24,7 @@ namespace Friendica\Test\src\Core\Lock;
 use Exception;
 use Friendica\Core\Cache\Type\RedisCache;
 use Friendica\Core\Config\IConfig;
-use Friendica\Core\Lock\CacheLock;
+use Friendica\Core\Lock\Type\CacheLock;
 use Mockery;
 
 /**
@@ -62,7 +62,7 @@ class RedisCacheLockTest extends LockTest
 
 		try {
 			$cache = new RedisCache($host, $configMock);
-			$lock = new CacheLock($cache);
+			$lock = new \Friendica\Core\Lock\Type\CacheLock($cache);
 		} catch (Exception $e) {
 			static::markTestSkipped('Redis is not available. Error: ' . $e->getMessage());
 		}
